Revert "Merge branch '2.6-beta' into 2.5-beta"
This reverts commit 8e78d1e728fe067ff72c75079e54c8f08664d70c, reversing
changes made to 5085603455941d4316ca840726400a4e2f14caf1.

+++ server/graphql/types/order-inventory/order-inventory-patch.ts
@@ -15,7 +15,7 @@ export const OrderInventoryPatch = gql`
     productName: String
     batchId: String
     packingType: String
-    releaseQty: Float
+    releaseQty: Int
     releaseWeight: Float
     remark: String
     issue: String
